+++ libharp/harp-area-mask.c
@@ -241,7 +241,7 @@ static int parse_polygon(const char *str, harp_spherical_polygon **polygon)
             str++;
         }
 
-        length = str - mark;
+        length = (int)(str - mark);
         if (harp_parse_double(mark, length, &point.lat, 0) != length || !harp_isfinite(point.lat))
         {
             harp_set_error(HARP_ERROR_INVALID_FORMAT, "invalid latitude '%.*s' (%s:%u)", length, mark, __FILE__,
@@ -275,7 +275,7 @@ static int parse_polygon(const char *str, harp_spherical_polygon **polygon)
             str++;
         }
 
-        length = str - mark;
+        length = (int)(str - mark);
         if (harp_parse_double(mark, length, &point.lon, 0) != length || !harp_isfinite(point.lon))
         {
             harp_set_error(HARP_ERROR_INVALID_FORMAT, "invalid longitude '%.*s' (%s:%u)", length, mark, __FILE__,
