+++ libharp/harp-ingest-qa4ecv.c
@@ -1006,7 +1006,7 @@ static int read_no2_column_stratospheric_avk(void *user_data, harp_array data)
         {
             for (j = 0; j < info->num_layers; j++)
             {
-                data.float_data[i * info->num_layers + j] = harp_nan();
+                data.float_data[i * info->num_layers + j] = (float)harp_nan();
             }
         }
         else
@@ -1111,7 +1111,7 @@ static int read_no2_column_tropospheric_avk(void *user_data, harp_array data)
         {
             for (j = 0; j < info->num_layers; j++)
             {
-                data.float_data[i * info->num_layers + j] = harp_nan();
+                data.float_data[i * info->num_layers + j] = (float)harp_nan();
             }
         }
         else
