+++ libharp/harp-ingestion-module.c
@@ -1093,7 +1093,7 @@ void harp_product_definition_add_mapping(harp_product_definition *product_defini
             long length;
 
             /* append description */
-            length = strlen(product_definition->mapping_description) + strlen(mapping_description) + 1;
+            length = (long)strlen(product_definition->mapping_description) + strlen(mapping_description) + 1;
             new_description = realloc(product_definition->mapping_description, length);
             assert(new_description != NULL);
             strcat(new_description, mapping_description);
